Moved the showFailureSnackBar function to a seperate folder under the widgets dir
