+++ packages/clima_ui/lib/screens/loading_screen.dart
@@ -4,7 +4,7 @@ import 'package:clima_ui/main.dart';
 import 'package:clima_ui/state_notifiers/forecasts_state_notifier.dart' as f;
 import 'package:clima_ui/state_notifiers/weather_state_notifier.dart';
 import 'package:clima_ui/utilities/hooks.dart';
-import 'package:clima_ui/utilities/reusable_widgets.dart';
+import 'package:clima_ui/widgets/failure_snack_bar.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
